Store GLO Limits as Double Precision on Restart
That way we don't impose a second double->float conversion after
float->double for unit conversion.  This enables reaching the
maximum lift gas injection rate in an example model, but is
nevertheless at best a stop-gap measure.  GLO results in restarted
runs typically do not yet match those of the base run.

+++ opm/io/eclipse/rst/group.hpp
@@ -68,8 +68,8 @@ struct RstGroup
     float gas_reservoir_limit;
     float gas_reinject_limit;
     float gas_voidage_limit;
-    float glift_max_supply;
-    float glift_max_rate;
+    double glift_max_supply;
+    double glift_max_rate;
     float efficiency_factor;
     float inj_water_guide_rate;
     float inj_gas_guide_rate;
